rename classes to avoid using Servlet in the name as it is confusing #2370
