+++ jetty-maven-plugin/src/it/jetty-run-distro-mojo-it/jetty-simple-webapp/src/test/java/org/eclipse/jetty/its/jetty_run_distro_mojo_it/TestGetContent.java
@@ -16,8 +16,7 @@
 //  ========================================================================
 //
 
-
-package org.eclipse.jetty.its.jetty_run_war_exploded_mojo_it;
+package org.eclipse.jetty.its.jetty_run_distro_mojo_it;
 
 import java.io.File;
 import java.io.FileReader;
@@ -30,10 +29,11 @@ import org.junit.Test;
 /**
  *
  */
-public class TestHelloServlet
+public class TestGetContent
 {
+
     @Test
-    public void hello_servlet()
+    public void get_ping_response()
         throws Exception
     {
         int port = getPort();
