+++ jetty-maven-plugin/src/it/jetty-run-mojo-it/jetty-simple-webapp/src/test/java/org/eclipse/jetty/its/jetty_run_mojo_it/TestGetContent.java
@@ -29,10 +29,11 @@ import org.junit.Test;
 /**
  *
  */
-public class TestHelloServlet
+public class TestGetContent
+
 {
     @Test
-    public void hello_servlet()
+    public void get_ping_response()
         throws Exception
     {
